requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=1.2.0
aiofiles>=23.0.0
python-multipart>=0.0.6
anthropic>=0.19.0

# 3D Model Generation
trimesh>=4.0.0
numpy-stl>=3.0.0
Pillow>=10.1.0
pygltflib>=1.16.0